Clean method implemented.
