+++ src/archive/clean.rs
@@ -1,7 +1,6 @@
 //! The cleaning method for Archive is complex, so it has its own module.
 
-/*
-use std::{collections::HashSet, convert::TryFrom, io::Read, str::FromStr};
+use std::{collections::HashSet, io::Read, str::FromStr};
 
 use crate::{
     coords::Coords,
@@ -10,11 +9,10 @@ use crate::{
     site::{SiteInfo, StationNumber},
 };
 
-use super::Archive;
-*/
+use metfor::{Meters, Quantity};
 
-/*
-// FIXME: Completely redo this.
+use super::Archive;
+
 impl Archive {
     /// Validate files listed in the index are in the archive too, if not remove them from the
     /// index.
@@ -40,9 +38,6 @@ impl Archive {
         println!("Comparing sets for files in archive but not in the index.");
         let files_not_in_index = file_system_vals.difference(&index_vals);
         self.handle_files_in_archive_but_not_index(&arch, &mut files_not_in_index.into_iter())?;
-
-        println!("Checking for orphaned stations.");
-        self.handle_orphaned_stations(&arch)?;
 
         println!("Compressing index.");
         arch.db_conn.execute("VACUUM", rusqlite::NO_PARAMS)?;
@@ -104,30 +99,48 @@ impl Archive {
     ) -> Result<(), BufkitDataErr> {
         let mut insert_stmt = arch.db_conn.prepare(
             "
-                INSERT INTO files (station_num, model, init_time, end_time, file_name)
-                VALUES (?1, ?2, ?3, ?4, ?5)
+                INSERT INTO files (
+                    station_num, 
+                    model,
+                    init_time,
+                    end_time,
+                    file_name, 
+                    id, 
+                    lat, 
+                    lon, 
+                    elevation_m
+                )
+                VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)
             ",
         )?;
 
         arch.db_conn
             .execute("BEGIN TRANSACTION", rusqlite::NO_PARAMS)?;
         for extra_file in files_not_in_index {
-            let message = if let Some((init_time, end_time, model, site, coords)) =
+            let message = if let Some((station_num, model, id, init, end, coords, elev)) =
                 arch.extract_site_info_from_file(&extra_file)
             {
-                let site: Site = if let Some(site_found) = arch.site(site.station_num) {
-                    site_found
-                } else {
-                    arch.add_site(&site, coords)?;
-                    site
+                if arch.site(station_num).is_none() {
+                    let site = SiteInfo {
+                        station_num,
+                        ..SiteInfo::default()
+                    };
+
+                    arch.add_site(&site)?;
                 };
 
+                let station_num: u32 = station_num.into();
+
                 match insert_stmt.execute(&[
-                    &site.station_num,
+                    &station_num as &dyn rusqlite::types::ToSql,
                     &model.as_static_str() as &dyn rusqlite::types::ToSql,
-                    &init_time as &dyn rusqlite::types::ToSql,
-                    &end_time as &dyn rusqlite::types::ToSql,
+                    &init as &dyn rusqlite::types::ToSql,
+                    &end as &dyn rusqlite::types::ToSql,
                     &extra_file,
+                    &id,
+                    &coords.lat,
+                    &coords.lon,
+                    &elev.unpack(),
                 ]) {
                     Ok(_) => format!("Added {}", extra_file),
                     Err(_) => {
@@ -149,45 +162,17 @@ impl Archive {
         Ok(())
     }
 
-    #[inline]
-    fn handle_orphaned_stations(&self, arch: &Archive) -> Result<(), BufkitDataErr> {
-        let mut stations_with_ids_stmt =
-            arch.db_conn.prepare("SELECT station_num FROM site_ids")?;
-        let stations_with_ids: Result<HashSet<i64>, BufkitDataErr> = stations_with_ids_stmt
-            .query_map(rusqlite::NO_PARAMS, |row| row.get::<_, i64>(0))?
-            .map(|res| res.map_err(BufkitDataErr::Database))
-            .collect();
-        let stations_with_ids = stations_with_ids?;
-        let mut stations_in_index_stmt = arch
-            .db_conn
-            .prepare("SELECT DISTINCT station_num FROM sites")?;
-        let stations_in_index: Result<HashSet<i64>, BufkitDataErr> = stations_in_index_stmt
-            .query_map(rusqlite::NO_PARAMS, |row| row.get::<_, i64>(0))?
-            .map(|res| res.map_err(BufkitDataErr::Database))
-            .collect();
-        let stations_in_index = stations_in_index?;
-
-        let orphans = stations_in_index.difference(&stations_with_ids);
-        for &orphan in orphans {
-            if let Some(site) = arch.site(orphan as u32) {
-                println!("     {}", site);
-            } else {
-                println!("     {} - unknown", orphan);
-            }
-        }
-
-        Ok(())
-    }
-
     fn extract_site_info_from_file(
         &self,
         fname: &str,
     ) -> Option<(
-        chrono::NaiveDateTime,
-        chrono::NaiveDateTime,
+        StationNumber,
         Model,
-        Site,
+        Option<String>,
+        chrono::NaiveDateTime,
+        chrono::NaiveDateTime,
         Coords,
+        Meters,
     )> {
         let tokens: Vec<&str> = fname.split(|c| c == '_' || c == '.').collect();
 
@@ -196,36 +181,29 @@ impl Archive {
         }
 
         let model = Model::from_str(tokens[1]).ok()?;
-        let id = Some(tokens[2].to_owned());
 
         let file = std::fs::File::open(self.data_root().join(fname)).ok()?;
         let mut decoder = flate2::read::GzDecoder::new(file);
         let mut s = String::new();
         decoder.read_to_string(&mut s).ok()?;
 
-        let snds = sounding_bufkit::BufkitData::init(&s, fname).ok()?;
-        let mut snds = snds.into_iter();
+        let (station_num, parsed_site_id, init_time, end_time, coords, elevation) =
+            Self::parse_site_info(&s).ok()?;
 
-        let first = snds.next()?.0;
-        let last = snds.last()?.0;
-
-        let init_time = first.valid_time()?;
-        let end_time = last.valid_time()?;
-
-        let station_num: u32 = first
-            .station_info()
-            .station_num()
-            .into_option()
-            .and_then(|int32| u32::try_from(int32).ok())?;
-
-        let coords: Coords = first.station_info().location().map(Coords::from)?;
-
-        let site = Site {
-            station_num,
-            id,
-            ..Site::default()
+        let id = if parsed_site_id.is_some() {
+            parsed_site_id
+        } else {
+            Some(tokens[2].to_owned())
         };
-        Some((init_time, end_time, model, site, coords))
+
+        Some((
+            station_num,
+            model,
+            id,
+            init_time,
+            end_time,
+            coords,
+            elevation,
+        ))
     }
 }
-*/
